refactor(posit): split posit into core / io / debug layers (#1334 pilot)
The Phase 1 pilot. posit is now three headers, and posit.hpp composes all three,
so nothing existing changes:

    #include <universal/number/posit/posit.hpp>   // everything, as before
    #include <universal/number/posit/core.hpp>    // arithmetic only

posit.hpp orders them core -> manipulators -> iostream -> debug.

The text layer goes where the library ALREADY KEEPS IT. manipulators.hpp is the
<iomanip> half -- to_binary, to_triple, to_native, to_string, hex_format,
to_base2_scientific, quadrant and parse join type_tag, hex_print, components,
pretty_print and color_print, which were there already. CLAUDE.md documents
manipulators.hpp as the home for to_binary/to_hex, and lns and dd already follow
it; cfloat, fixpnt, integer and takum leave theirs in _impl.hpp, which is the
inconsistency this epic should be removing rather than adding to.

iostream.hpp is the <iostream> half and holds only what genuinely needs the stream
types: operator<< and operator>>. 93 lines.

THE FACET HEADERS ARE NAMED AFTER WHAT THEY MIRROR. Every number type already
carries the same unprefixed set -- attributes.hpp, exceptions.hpp,
manipulators.hpp, mathlib.hpp, numeric_limits.hpp -- while the TYPE_-prefixed
names in the same directory are implementation internals (posit_impl, posit_fwd,
posit_regime, posit_fraction, posit_exponent, posit_scale_helpers). And
numeric_limits.hpp is already named for the standard header it mirrors, so
iostream.hpp is the same move rather than a new idea. My first cut called these
posit_io.hpp and posit_debug.hpp, which matched neither convention; they are
iostream.hpp and debug.hpp.

parse() sits with the manipulators rather than with the streams: it is string ->
posit, the inverse of to_string, and touches no stream.

ulp() stayed in the core -- it sat inside the region that moved but is arithmetic,
not text.

debug.hpp takes constexprClassParameters() and showLimbs() as out-of-line
member definitions, the same shape #1388 used for blocktriple.

The 18 trace sites are wrapped in #if POSIT_TRACE_ENABLED, keyed on the existing
ALGORITHM_TRACE_* macros, with <iostream>/<iomanip> conditional on the same
switch. `if constexpr` cannot serve here for the reason this epic keeps meeting: a
discarded `if constexpr` branch still requires std::cout to be DECLARED, so the
include would stay in every translation unit's graph.

posit_impl.hpp is left with zero std::cout, zero std::stringstream, zero
std::string, and <iosfwd> for the two friend declarations.

THE TARGET IS NOT MET, AND THE REASON IS WORTH THE PILOT.

    posit.hpp   114656 lines   497 headers   5 I/O-family
    core.hpp    107524 lines   446 headers   5 I/O-family   -6.2%

against a target of under 45000 and zero I/O-family headers. posit's own code is
clean; its DEPENDENCIES are not:

    iomanip, sstream  <- internal/blockbinary/blockbinary.hpp (its own to_binary/to_string)
    iostream          <- number/integer/integer.hpp, reached through
                         posit_impl -> utility/decimal_to_binary.hpp

posit's core pulls the entire integer type because decimal conversion needs
arbitrary-precision integers. Reaching the target needs this same split applied to
blockbinary and integer -- the epic's "layering cost" phase, which this measures
as required rather than optional. That is the finding a pilot exists to produce:
the per-type split is necessary and not sufficient.

Verified: 38/38 umbrellas on gcc 13.3 and clang 18.1; core.hpp compiles and
computes on its own; the posit test suites build and pass; CI_LITE builds and
passes.

Relates to #1334
